feat: application Setup 함수를 하나로 합쳐 일원화
