+++ src/main.ts
@@ -3,7 +3,6 @@ import { NestFactory } from '@nestjs/core';
 import type { NestExpressApplication } from '@nestjs/platform-express';
 
 import { winstonLogger } from '#configs/logger.config';
-import { setupSwaggerModule } from '#configs/swagger.config';
 
 import { AppModule } from './app.module';
 import { setupNestApplication } from './common/configs/setup.config';
@@ -14,11 +13,10 @@ async function bootstrap() {
 		logger: winstonLogger,
 	});
 
+	setupNestApplication(app);
+
 	const configService = app.get(ConfigService);
 	const port = configService.get<number>('SERVER_PORT') || 3000;
-
-	setupNestApplication(app);
-	setupSwaggerModule(app);
 
 	await app.listen(port);
 }
